pyproject.toml
[project]
name = "weatherstationdatabridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alessandro Annini", email = "alessandro.annini@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "typer[all]>=0.19.2",
]

[project.scripts]
weatherstationdatabridge = "weatherstationdatabridge:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
]
